# internal/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_krkn-chaos_krkn-operator_api_v1alpha1.ScenarioReference:
    properties:
      name:
        description: Name is the scenario tag/name to resolve.
        type: string
      private:
        description: |-
          Private selects a saved private registry when true, or krknctl's public
          Quay provider when false. A pointer makes the field mandatory on input.
        type: boolean
      registryName:
        description: RegistryName identifies the saved private registry when Private
          is true.
        type: string
    type: object
...
  github_com_krkn-chaos_krkn-operator_pkg_files.FileReference:
    properties:
      fileId:
        description: FileID is the UUID of the file to mount
        type: string
      mountPath:
        description: MountPath is the absolute path where the file should be mounted
        type: string
    type: object
...
  internal_api.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  internal_api.FileMount:
    properties:
      content:
        description: Content is the base64-encoded file content
        type: string
      mountPath:
        description: MountPath is the absolute path where the file should be mounted
        type: string
      name:
        description: Name is the file name
        type: string
    type: object
...
  internal_api.ScenarioRunRequest:
    properties:
      categories:
        description: Categories are visible category names to associate with the run
          when it is created.
        items:
          type: string
        type: array
      cloudCredentialRef:
        description: |-
          CloudCredentialRef, if set, names a saved cloud credential Secret whose
          reference is set on the CRD spec for controller-level SecretKeyRef injection.
        type: string
      customRunName:
        description: CustomRunName is a user-provided label for the run (optional)
        type: string
      elasticsearchConfigName:
        description: |-
          ElasticsearchConfigName, if set, names a saved Elasticsearch config Secret whose
          credentials (ES_PASSWORD, and any ES_* vars not already in Environment) are
          injected server-side so the password is never transmitted by the client.
        type: string
      environment:
        additionalProperties:
          type: string
        description: Environment is a map of environment variables to pass to the
          container (optional)
        type: object
      fileReferences:
        description: FileReferences are references to centrally-managed files by UUID
          (optional)
        items:
          $ref: '#/definitions/github_com_krkn-chaos_krkn-operator_pkg_files.FileReference'
        type: array
      files:
        description: Files is an array of file objects to mount in the container (optional,
          legacy inline file mount)
        items:
          $ref: '#/definitions/internal_api.FileMount'
        type: array
      kubeconfigPath:
        description: 'KubeconfigPath is the path where kubeconfig should be mounted
          (optional, default: /home/krkn/.kube/config)'
        type: string
      registryName:
        description: |-
          RegistryName is the name of a saved registry (optional)
          If omitted, defaults to quay.io public registry
        type: string
      resiliencyScoreEnabled:
        description: |-
          ResiliencyScoreEnabled enables resiliency score calculation for this scenario run.
          When true, RESILIENCY_SCORE=true is set in scenario pods and the controller
          calculates scores from pod logs upon completion.
        type: boolean
      scenario:
        allOf:
        - $ref: '#/definitions/github_com_krkn-chaos_krkn-operator_api_v1alpha1.ScenarioReference'
        description: |-
          Scenario identifies the scenario and registry to resolve. The operator
          resolves the executable image through krknctl; callers cannot provide one.
      scenarioImage:
        description: |-
          ScenarioImage is accepted for backwards compatibility but is ignored.
          Images are always resolved server-side from Scenario.
          ScenarioImage is always serialized for config responses. It is empty
          until the controller has resolved an executable image for a job.
        type: string
      scenarioName:
        description: |-
          ScenarioName is the legacy scenario identity field. It is translated to
          Scenario when a request does not include the new reference object.
        type: string
      targetClusters:
        additionalProperties:
          items:
            type: string
          type: array
        description: |-
          TargetClusters is a map of provider-name to list of cluster names
          Example: {"krkn-operator": ["cluster1", "cluster2"], "krkn-operator-acm": ["cluster3"]}
        type: object
      targetRequestId:
        description: TargetRequestID is the UUID of the KrknTargetRequest (required)
        type: string
    type: object
  internal_api.ScenarioTag:
    properties:
      digest:
        description: Digest is the image digest (optional)
        type: string
      lastModified:
        description: LastModified is when the scenario was last updated (optional)
        type: string
      name:
        description: Name is the scenario tag/version name
        type: string
      signature_status:
        description: 'SignatureStatus is the image signature state: signed, unsigned,
          or unknown.'
        type: string
      size:
        description: Size is the image size in bytes (optional)
        type: integer
    type: object
  internal_api.ScenariosResponse:
    properties:
      scenarios:
        description: Scenarios contains the list of available scenario tags
        items:
          $ref: '#/definitions/internal_api.ScenarioTag'
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: krkn-chaos@googlegroups.com
    name: Krkn Team
    url: https://github.com/krkn-chaos/krkn-operator
  description: |-
    REST and WebSocket API for Krkn chaos engineering operator.

    **API Versions:**
    - **v1** - REST API with polling (deprecated but maintained)
    - **v2** - REST API with v2-specific resources + WebSocket real-time updates

    **WebSocket Authentication (v2):**
    WebSocket endpoints use JWT via subprotocol header:
    - JavaScript: `new WebSocket(url, 'access_token.' + jwtToken)`
    - Header: `Sec-WebSocket-Protocol: access_token.<jwt_token>`

    **Migration Path:**
    1. v1 REST → v2 REST (no changes, just update base path)
    2. v2 REST → v2 WebSocket (replace polling with multiplexed WebSocket)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://krkn-chaos.dev/terms
  title: Krkn Operator API
  version: "2.0"
paths:
...
  /scenarios:
    post:
      consumes:
      - application/json
      description: Get list of available chaos scenarios from container registry (Quay.io
        or private registry)
      parameters:
      - description: Registry configuration (optional for private registries)
        in: body
        name: registry
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: List of available scenarios
          schema:
            $ref: '#/definitions/internal_api.ScenariosResponse'
        "400":
          description: Invalid registry configuration
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List available scenarios
      tags:
      - scenarios
...
  /scenarios/run/replay/{jobId}:
    get:
      description: Retrieve scenario configuration from a completed job and return
        payload ready for re-execution via POST /scenarios/run
      parameters:
      - description: Job ID (krkn-job-id label value)
        in: path
        name: jobId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Scenario configuration ready for replay
          schema:
            $ref: '#/definitions/internal_api.ScenarioRunRequest'
        "400":
          description: Invalid job ID or job not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: Job or ScenarioRun not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Replay a scenario from a completed job
      tags:
      - scenarios
...
securityDefinitions:
  BearerAuth:
    description: 'JWT token obtained from /api/v1/auth/login or /api/v2/auth/login.
      Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"